requirements.txt
asgiref==3.4.1
coloredlogs==15.0.1
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==9.3.1
humanfriendly==10.0
importlib-metadata==4.8.1
L7 @ file:///Users/kenedy/project/L7%2B/L7/L7_database/dist/L7-0.1.tar.gz
Markdown==3.3.4
nanoid==2.0.0
Pillow==8.4.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
zipp==3.6.0
